chore: deleted docker stuffs I decided to just deploy it on vercel lol.. initially I wanted to tryout to deploy my project on a VPS.. well I have tried it and it works..
But when I stumbled accross middleware stuffs, I noticed that if I use middleware on my nextjs and deploy it to vercel, my middleware would be ran into vercel's cdn and thus running on the edge. Well, if i use a VPS as to host my app, it would be quite hard to implement edge run time stuffs, to much of a hassle! my brain is crying..

so, vercel it is.. only for web hosting tho..